requirements.txt
plotly==4.6.0
python-dotenv==0.13.0
requests==2.23.0